feat(graphics): increased precision for raycasting calculations, as well as modifiable FOV calculation
